+++ relay/channel/ali/adaptor_test.go
@@ -1,0 +1,11 @@
+package ali
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestAdaptorDoesNotSupportAudioSpeech(t *testing.T) {
+	require.False(t, (&Adaptor{}).SupportsAudioSpeech())
+}